Small refactors

+++ crates/lgn-animation/src/animation_system.rs
@@ -11,7 +11,7 @@ pub(crate) fn graph_update(mut graphs: Query<'_, '_, &mut GraphDefinition>, time
         // Todo!
 
         // update the current node
-        (*graph.nodes[current_node_index as usize]).update(delta_time);
+        (*graph.nodes[current_node_index]).update(delta_time);
     }
     drop(graphs);
     drop(time);
